Refactored CSF codebase
