+++ examples/builder/blueprint/csf/__init__.py
@@ -1,2 +1,8 @@
-from .csf_manager import Affordance, Perception, Context, Brain, Interpreter
 from .utils import print_context
+
+from . import core
+from . import interpreters
+from . import frames
+from . import practices
+
+from .csf_manager import Brain
