+++ examples/builder/blueprint/csf/frames/__init__.py
@@ -1,1 +1,3 @@
 from .cognitive_social_frame import CognitiveSocialFrame
+from .human_frame import HumanFrame
+from .work_frame import WorkFrame
