+++ examples/builder/blueprint/csf/frames/cognitive_social_frame.py
@@ -1,12 +1,14 @@
 from abc import abstractclassmethod, ABC
 
+import csf.core
+
 
 class CognitiveSocialFrame(ABC):
     def __init__(self, perception_labels: set[str]) -> None:
         self._needed_labels = perception_labels
 
     @abstractclassmethod
-    def is_salient(self, context) -> bool:
+    def is_salient(self, context: csf.core.Context) -> bool:
         pass
 
     @abstractclassmethod
