+++ examples/builder/blueprint/csf/frames/human_frame.py
@@ -1,0 +1,22 @@
+from javascript import require
+
+from .cognitive_social_frame import CognitiveSocialFrame
+import csf.core
+import csf.practices
+
+
+Vec3 = require("vec3")
+pathfinder = require("mineflayer-pathfinder")
+
+class HumanFrame(CognitiveSocialFrame):
+
+    pass
+
+
+class Sleep(csf.practices.Practice):
+
+    pass
+
+
+class Eat(csf.practices.Practice):
+    pass
