+++ examples/builder/blueprint/csf/frames/work_frame.py
@@ -1,0 +1,17 @@
+import time
+
+import csf.practices
+from javascript import require
+
+from .cognitive_social_frame import CognitiveSocialFrame
+
+Vec3 = require("vec3")
+pathfinder = require("mineflayer-pathfinder")
+
+
+class WorkFrame(CognitiveSocialFrame):
+    pass
+
+
+class GoToWork(csf.practices.Practice):
+    pass
