+++ examples/builder/blueprint/csf/interpreters/interpreter.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+import csf.core
+
+
+class Interpreter(ABC):
+    @abstractmethod
+    def process_perceptions(self, perceptions: csf.core.Context) -> list[csf.core.Perception]:
+        pass
