+++ examples/builder/blueprint/csf/practices/__init__.py
@@ -1,0 +1,1 @@
+from .practice import Practice
